backend/pyproject.toml
[project]
name = "fetal-plane-explorer-backend"
version = "0.1.0"
description = "Backend service for Fetal Plane Explorer"
authors = [
    {name = "Olaf Laitinen", email = "olaf@example.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "python-multipart>=0.0.9",
    "onnxruntime>=1.17.0",
    "numpy>=1.26.0",
    "pillow>=10.3.0",
    "orjson>=3.10.0",
    "torch>=2.2.0",
    "torchvision>=0.17.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pytest-cov>=5.0.0",
    "types-pillow",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = []

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

